refactor: Move Category to enums and update references across the codebase
