+++ app/api/src/main/java/com/somba/api/adapter/controllers/GlobalExceptionHandler.java
