+++ app/api/src/main/java/com/somba/api/adapter/repositories/MdbProductRepositoryAdapter.java
@@ -4,8 +4,8 @@ import java.util.List;
 import org.springframework.data.domain.PageRequest;
 
 import com.somba.api.adapter.mappers.ProductMapper;
-import com.somba.api.core.entities.Category;
 import com.somba.api.core.entities.Product;
+import com.somba.api.core.enums.Category;
 import com.somba.api.core.ports.ProductRepository;
 import com.somba.api.infrastructure.persistence.MdbProductRepository;
