+++ app/api/src/main/java/com/somba/api/core/enums/Category.java
@@ -1,4 +1,4 @@
-package com.somba.api.core.entities;
+package com.somba.api.core.enums;
 
 import com.somba.api.core.exceptions.InvalidCategoryException;
 import com.somba.api.core.exceptions.NullCategoryException;
